KOGITO-3073 Create protobuf common module to allow reusing marshallers outside Infinispan addon (#811)
