+++ addons/persistence/infinispan-persistence-addon/src/test/java/org/kie/kogito/infinispan/KogitoProcessInstancesFactoryTest.java
@@ -26,9 +26,9 @@ import org.jbpm.workflow.core.impl.WorkflowProcessImpl;
 import org.junit.jupiter.api.Test;
 import org.kie.api.runtime.process.WorkflowProcessInstance;
 import org.kie.kogito.Model;
-import org.kie.kogito.infinispan.marshallers.BooleanMessageMarshaller;
-import org.kie.kogito.infinispan.marshallers.StringMessageMarshaller;
 import org.kie.kogito.persistence.KogitoProcessInstancesFactory;
+import org.kie.kogito.persistence.protobuf.marshallers.BooleanMessageMarshaller;
+import org.kie.kogito.persistence.protobuf.marshallers.StringMessageMarshaller;
 import org.kie.kogito.process.Process;
 import org.kie.kogito.process.ProcessInstance;
 import org.kie.kogito.process.impl.AbstractProcess;
