+++ addons/persistence/protobuf-persistence-common/src/main/resources/kogito-types.proto
